+++ packages/v4/src/content/design-guidelines/styles/colors/ColorFamily.css
@@ -4,15 +4,21 @@
   margin-right: var(--pf-c-accordion__toggle--PaddingLeft)
 }
 
-.ws-color-family-color {
-  font-family: var(--pf-global--FontFamily--monospace);
-}
-
 /* Specificity YAY */
 .pf-c-accordion__toggle.ws-color-family-content {
   font-size: var(--pf-global--FontSize--xs);
   display: inline-block;
   padding-bottom: var(--pf-global--spacer--lg);
+}
+
+/* Specificity YAY */
+.pf-c-accordion__toggle.ws-color-family-content:hover,
+.pf-c-accordion__toggle.ws-color-family-accordion-toggle:hover {
+  background-color: unset;
+}
+
+.ws-color-family-accordion-toggle {
+  cursor: pointer;
 }
 
 .pf-c-accordion__toggle > .ws-color-family-label {
@@ -23,7 +29,9 @@
   margin-bottom: var(--pf-global--spacer--sm);
 }
 
-.ws-color-family-toggle.pf-c-accordion__toggle::before {
+/* Specificity YAY */
+.ws-color-family-toggle.pf-c-accordion__toggle::before,
+.ws-color-family-accordion-toggle.pf-c-accordion__toggle::before {
   content: none;
   width: 0;
 }
